requirements.txt
# packaging
pip>=9
setuptools>=26
twine
wheel>=0.29

# data science
numpy>=1.18.5
pandas>=1.4.1
scikit-learn
matplotlib
seaborn
empiricaldist

# tests/linter
black
coverage
flake8
pytest
yapf

# API
gcsfs
google-cloud-storage
mlflow
bs4

# utilities
six>=1.14
joblib
memoized-property
termcolor
python-dotenv
regex

# db
sqlalchemy
pymssql
